Temporary stack variable should be the type of its previous object
In certain situations the stack of tmp_var's had the wrong type associated with it, as each replacement entry was being created with the type of the expression it was built from. In this particular case a java array created and directly passed into a method was causing the 'this' temp variable to be stored as the type of the array itself, causing an assertion error further on.

The type of code causing this error:

this.objectInstance.Method(new int[]{0, 1, 2});