Remove unused type temp_working_dirt
This type has been added in 2012 and is presently unused. In its present state
it is difficult to use as it doesn't permit any way to recover from potential
errors (it is instead asserted that no errors can happen, which is not true).

Should functionality like this be needed in the future, I'd recommend a wrapper
function like

```
template<typename Closure>
void run_with_different_working_dir(
  const std::string& working_dir, Closure closure);
```

Which would change the working directory, execute the closure and change the
working directory back afterwards, throwing a system_exceptiont on failure.