Update comment to make doxygen happy
Doxygen produced a warning about a reference it couldn't find, because
of a '#' symbol used in the comment (which the author didn't intend to
be a reference).  Usually, '#' could be escaped '\#', but that doesn't
work in this case, because it appears within quotes (as explained in
https://stackoverflow.com/a/36750674 ).  I couldn't find a way to get
this to work with the quotes that had been used, so I have changed it
to use different quotes.