Make string_not_contains_constraintt a struct
In particular it does not inherit from exprt anymore.
Not inheriting from exprt, allows the class to be more structured.
We have fields that have names describing what they contain, instead of
having to define accessor methods.
This also prevent us from mistakenly passing a constraint to an exprt
function, which wouldn't know how to deal with a
string_not_contains_constraintt.