Allocate nondet char array
make_nondet_infonite_char_array was not allocating the array as a
dynamic object. This means if this code was added in a function called
several time then the array would get overwritten each time.
This problem was not visible by the string refinement as it associates a
independent array to each string and does not care about memory
allocation, but can be a problem for the interpreter.