Remove User Manual from Doxygen documentation
Remove the version of the user manual published at
http://cprover.diffblue.com/cbmc-user-manual.html
and link to http://www.cprover.org/cprover-manual/ instead, in order to
make a clearer separation between the _user_ and _developer_
documentation.

There is updated source intended for the user manual, with almost
identical contents to the removed cbmc-user-manual.md, in the
doc/cprover-manual folder.