Add high-level documentation about Java traces.
This overview is intended to give an overview of what a user might see if they inspect a trace of the execution of a Java program. It details function calls, function returns and gives examples assignments, but notin detail what the expressions on each side of the assignments look like.