Use range in symex instead of merging sets of expressions
This part of the code is executed in all our tests so the fact that all
existing tests pass should show this change works correctly.

The performance difference for this change were evaluate on 143 methods
from tika, which were chosen because analysis time was taking between 1
second and 1 minute for them.
The time of symex step was evaluated alone (options --show-vcc --verbosity 0).

The execution time for the whole of the benchmarks was reduced by 21%
and the average speed up is 14% (the difference between the two means
the optimization seems to have less influence on simpler benchmarks).