Treat unexpected Doxygen output as warnings
Change the Doxygen output filter so that the Travis job fails when it
encounters something unexpected.

Previously, it was printing a message to the log when it encountered
something unexpected, but not changing the return code.  For example,
around 2018-11-14 (see e.g. commit ca06683b6), the Travis job succeeded
because this warning was not recognised:

  <unknown>:97: warning: unable to resolve reference to
  `instrumentation' for \ref command