Mark return type as needed on orphaned virtual call
In a situation where we've seen a virtual call but can't work out where that call may come from we need to assert that the return type has also been 'seen' even if its' not initalized anywhere else. In the particular instance I was coming up against it was an abstract method whose return type had an enum field, and to nondet initalize that return type it needed the enum clinit method to exists, and without this type being marked as seen that method was never generated.