Adding destructor_treet
This uses a grapht under the hood to keep track of nodes and just builds them in a tree shape. We don't use the usual graph algorithms for searching because they proved a little too costly when it was originally tried, and because we know what form the data will be in we can write very simple walking algorithms that take advantage of that.