Replacing the destructor stack with a tree structure
Copying the destructor stack at every instance of a goto / label gets more costly the larger the stack is, and in some niche situations was the bottleneck in generating GOTO. The original instance where this issue was seen the destructor stack got into the hundreds of thousands and took a significant amount of time to copy even once. While this is an unrealistic scenario, the change will just be faster in general while also allowing for better scaling.

On a simulated test of 40,000 labels / GOTOs in a single method it originally took 5 minutes to analyze, cut down to 30~ seconds with this change.

Memory usage is also better due to not copying everything around, though I didn't measure how much of a benefit it was.