Update `map_iteratort` to use a `shared_ptr` for `current`
This allows mapping into move only types. This was not previously
possible, because the copy constructor of `map_iteratort` copied the
value pointed to by its unique pointer in its `current` field. This can
also be considered to be more correct, because copying an iterator
should be expected to yield two copies, which point to the same value,
rather than two copies, which point to two separate values.