Remove no-longer-used system_symbols
This is a follow-up to the cleanup done in 80c9c85ed8315. Creating a
system_symbol_libraryt every time a languaget is created (effectively everytime
from_expr is called) is costly, because of all the strings being looked up in
the string table. On SV-COMP benchmarks, this accounted for 2% of runtime.
Furthermore this removes this undesirable dependency on goto-programs from
langapi.