replace incomplete_struct/_union/_c_enum by a flag in struct_/union_/c_enum_type
1) The incomplete_struct type is replaced by the struct type with ID_incomplete set.

2) The incomplete_union type is replaced by the union type with ID_incomplete set.

3) The incomplete_c_enum type is replaced by the c_enum type with ID_incomplete set.

The question whether any of the compound types is incomplete is only
relevant to the front-ends; outside of the front-end, the existence of three
further types only adds unnecessary complexity.

This change further establishes the invariant that a struct_tag, union_tag,
c_enum_tag always points to a struct, union, c_enum type, respectively.