Add tag to synthetic type made for String.format()
Having this type without a tag made it very hard to identify,
e.g. if we want to treat it differently from types which
are not synthetic (i.e. which come directly from the source
code). Compare with refined_string_typet, which has the tag
CPROVER_PREFIX "refined_string_type".