Ensure member_exprt's type matches the addressed component
Previously we created the expression with the partial type given at the
access site (i.e. as part of the
get/setfield bytecode), but this lacked any generic type information,
which the actual component *did* have.
The member expression could therefore be type-inconsistent with the
object is addressed. Since it's no longer
necessary to make the expression approximately and later fix it up in
typecheck_expr, let's just get it right
the first time, including giving it the right type.