Remove `objectt::` specifiers in types in `json_objectt`
Becuase `const_iterator` in `json_object` is defined to be
`objectt::const_iterator` and `iterator` is defined to be
`objectt::iterator`, this commit can remove occurrences of `objectt::`
and keep the actual types the same. The intention of this commit is to
improve readability, by reducing the amount of noise in these type
definitions.