Fix constness on dereferencing a `filter_iteratort`
Dereferencing a `filter_iteratort` from filtering a const collection
with `ranget` would previously introduce a compile error, due to a
`const` to non-const conversion. This commit fixes this, by using the
same return type as the iterator which it wraps around.

This commit includes a test for using range to filter a `const`
collection.