Rename `ranget::value_typet` to `ranget::value_type`
Algorithms which are written expecting a STL container may reference a
`value_type` member, as this is the name of the member in STL
containers. Therefore this needs to be named the same way in `ranget` in
order to be compatible with such algorithms.

This commit includes a unit test of `ranget` having correct the
`value_type` defined.