Add multi path symex only checker
multi_path_symex_only_checkert is a
bounded model checking algorithm that determines the
status of properties through symbolic execution and
constant propagation.
I.e. it doesn't call the SAT solver.

If desired CBMC could expose this to support a symex-only
BMC algorithm.

For now, this is used to implement show-vcc and program-only
(which are rather orthogonal to the BMC algorithm used,
but currently assume symex-only).