Add cast operator from `ranget` to a container
Adding `.collect` to the end of a chain of range operations can add
clutter to the code, rather than improving readability, in places where
the type of container being constructed is apparent from the context.
This commit adds a cast operation in order to improve readability in
this scenario.

In addition to the operator, this commit includes a refactor of
`optionst::to_json` in order to demonstrate an example usage of the
proposed change.