enable .clang-format-ignore
We do have files that are imported as a whole, and clang-formatting them is
a distraction.  The requirement to add // clang-format-off to external files
makes imports via scripts more difficult.

This tweaks the build script to exclude files given in .clang-format-ignore
from git-clang-format.  The same naming convention is used by other
projects.