Better goto-instrument error reporting for missing files
In the past, goto-instrument was failing with the default
--help error message if it didn't get the correct number
of positional arguments, making it unnecessarily hard to
understand what was failing and why (since the help message
is long and it also didn't have any particular signaling for
what went wrong). This makes it so that it reports it as an
invalid user input exception. Fixes #2861.