Make rename take a copy instead of reference
This makes the interface clearer as there is no argument that is both an
input and an output.
This also corresponds to how the function was used in a lot of
occurences, in which case the code is now simpler.
In the cases where we actually want the transformation to happen
in-place, we can use std::move.